repo_results/run-20240308-125808_LLMatic-ZeroShot-CodeT-SelfRefine-Reflexion-ZeroShot_donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-twitter-donnemartin-system-design-oop-whatsapp_0_20-runs_log/twitter/12/LLMatic/notification.py
class Notification:
	def __init__(self, id, user_id, post_id, type):
		self.id = id
		self.user_id = user_id
		self.post_id = post_id
		self.type = type

	def create_notification(self, user_id, post_id, type):
		self.id = len(Notification.notifications) + 1
		self.user_id = user_id
		self.post_id = post_id
		self.type = type
		Notification.notifications[self.id] = self

	@staticmethod
	def view_notifications(user_id):
		return [notification for notification in Notification.notifications.values() if notification.user_id == user_id]

	@staticmethod
	def clear_notifications():
		Notification.notifications.clear()

	notifications = {}
